Fix issues found by SonarCloud Remove unnecessary locks
